Split LL module upgrades (actions 1–7) into IGPX staging payload.
Moves UserModule/AdminModule upgrades and auth rotations off IGP-133
into contracts/payloads/IGPX for a separate governance proposal.
